# Faster generation of Sobol samples