--- a/src/Sobol.jl
+++ b/src/Sobol.jl
@@ -8,7 +8,24 @@ Base.@kwdef @concrete struct SobolSample <: DeterministicSamplingAlgorithm
 end
 
 function sample(n::Integer, d::Integer, S::SobolSample, T = Float64)
-    s = Sobol.SobolSeq(zeros(T, d), ones(T, d))
-    skip(s, n)
-    return randomize(reduce(hcat, [Sobol.next!(s) for i in 1:n]), S.R)
+    if n < 0
+        throw(ArgumentError("number of samples must be non-negative"))
+    end
+
+    seq = Matrix{T}(undef, d, n)
+    if n == 0
+        return seq
+    end
+
+    # Use function barrier since `Sobol.SobolSeq(d)` can't be inferred
+    return _sample!(seq, Sobol.SobolSeq(d), S.R)
+end
+
+function _sample!(seq::AbstractMatrix, s::Sobol.SobolSeq, R::RandomizationMethod)
+    n = size(seq, 2)
+    Sobol.skip!(s, n, @view(seq[:, begin]))
+    for x in eachcol(seq)
+        Sobol.next!(s, x)
+    end
+    return randomize(seq, R)
 end